Refactor: Move SplashActivity and configure Crashlytics
This commit refactors the project structure and introduces conditional configuration for Firebase Crashlytics and Ads.

Key changes:
- Moved `SplashActivity` to the `ui.splash` package to improve project organization.
- Created a custom `App` class to initialize Firebase Crashlytics.
- Added `ENABLE_CRASHLYTICS` and `ENABLE_ADS` boolean flags in `app/build.gradle` to control their activation in debug and release builds.
- Updated `AdsHelper.kt` to check the `ENABLE_ADS` flag before initializing or displaying ads.
- Modified `AndroidManifest.xml` to reference the new `App` class and the updated `SplashActivity` path.
