+++ app/src/main/res/layout/activity_splash.xml
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layoutDescription="@xml/activity_splash_scene"
-    tools:context="com.raylabs.doggie.SplashActivity">
+    tools:context="com.raylabs.doggie.ui.splash.SplashActivity">
 
     <androidx.appcompat.widget.AppCompatImageView
         android:id="@+id/appCompatImageView"
